# openapi.yaml
openapi: 3.1.0
info:
  title: NioDB API
  summary: NioDB — The Agentic DB that works.
  version: 0.1.0
  description: A lightweight database with natural-language queries, powered by Nio.
    Each Rust server has one workspace; clients do not supply workspace IDs. Nio CLI
    powers chat and read queries. AlaSQL executes restricted SQL. Recovery uses a
    checksummed JSONL journal, derived TOON artifacts and content-addressed file blobs.
    Backend credentials provision app users; users log in with username and password.
    Protected routes accept bearer credentials. Use HTTPS for remote access. Responses
    and errors carry X-Request-ID.
servers:
- url: /
  description: Same origin as the deployed NioDB server
security:
- bearerAuth: []
paths:
...
  /api/v1/storage/{bucket}:
    parameters:
    - &id003
      name: bucket
      in: path
      required: true
      schema:
        type: string
        pattern: ^[a-z0-9_-]{1,63}$
    get:
      operationId: listFiles
      tags:
      - Files
      summary: List bucket files
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - name: cursor
        in: query
        description: File ID returned as next_cursor, bound to this bucket.
        schema: &id004
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/File'
                  next_cursor:
                    type:
                    - string
                    - 'null'
        default: *id002
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer credential supporting combined client-token:secret-token (or individual client/secret tokens, both with full backend access), or expiring app-user session. Keep backend credentials off public web/mobile clients. Registration requires a backend credential; me/logout require a user session.
...
  schemas:
...
    File:
      type: object
      required:
      - id
      - record_id
      - artifact_id
      - bucket
      - filename
      - sha256
      - size
      - mime
      - created_at
      - created_by
      properties:
        id: *id004
        record_id: *id004
        artifact_id: *id004
        bucket: *id004
        filename: *id004
        sha256:
          type: string
          pattern: ^[a-f0-9]{64}$
        size:
          type: integer
          minimum: 0
          maximum: 16777216
        mime: *id004
        created_at:
          type: string
          format: date-time
        created_by: *id004
        user_id:
          type: string
          description: Present only for owner-linked files.
...